+++ sqlparser/test/utils/find_referenced_tables_test.dart
@@ -115,21 +115,5 @@ void main() {
     test('does not include references resolved in the query', () {
       testWith('WITH foo AS (VALUES(1,2,3)) SELECT * FROM foo', {});
     });
-
-    test('works for complex statements', () {
-      testWith('''
-CREATE TRIGGER my_trigger BEFORE DELETE on target BEGIN
-  SELECT * FROM old;
-  SELECT * FROM new; -- note: This is a delete trigger
-END
-''', {'target', 'new'});
-
-      testWith('''
-CREATE TRIGGER my_trigger BEFORE UPDATE on target BEGIN
-  SELECT * FROM old;
-  SELECT * FROM new;
-END
-''', {'target'});
-    });
   });
 }